lower/emit: promoted-method interface dispatch, variadic dispatch, InitLocals, shim sigs
The make-or-break fixes that take goja's compiled .dll from JIT-rejected to a valid,
loadable, ilverify-clean assembly:

- interface dispatch now handles a method PROMOTED from an embedded concrete field
  (goja's value types embed valueNull etc., so Export is the embedded type's). The
  case body navigates the embed field chain to the real receiver — dereferencing
  embedded pointers and boxing into a cell for a pointer-receiver promoted method —
  instead of passing the outer value, which the JIT rejected (InvalidProgramException
  / 'unexpected type on the stack').
- interface dispatch packs variadic args (m.Match() with no args now passes a nil
  slice), fixing a StackUnderflow.
- fat method header sets InitLocals (0x3013): CLR zero-inits locals (GC-safe, and
  unmasked the dispatch errors during ilverify).
- shim signatures matched to the Go-derived externs: atomic Int32 ops int (not
  long), reflect Value.Pointer ulong, MakeFunc takes GoClosure, time Ticker_C GoChan
  / Stop void, runtime.FuncForPC ulong.

goja now compiles to a 15 MB .dll, loads, JITs Main cleanly, and runs init. The
RunString eval path reaches execution and hits a runtime access violation (next).
conformance 356.

+++ runtime/stdlib/Goruntime.cs
@@ -10,11 +10,11 @@ public sealed class GoRuntimeFunc { public long Pc; }
 /// <summary>Shim for the subset of Go's <c>runtime</c> package that goja references.</summary>
 public static class Goruntime
 {
-    public static object FuncForPC(long pc) => new GoRuntimeFunc { Pc = pc };
+    public static object FuncForPC(ulong pc) => new GoRuntimeFunc { Pc = (long)pc };   // pc is uintptr
     public static GoString Func_Name(object f) => GoString.FromDotNetString("");
-    public static object?[] Func_FileLine(object f, long pc) =>
+    public static object?[] Func_FileLine(object f, ulong pc) =>
         new object?[] { GoString.FromDotNetString(""), 0L };
-    public static long Func_Entry(object f) => ((GoRuntimeFunc)f).Pc;
+    public static ulong Func_Entry(object f) => (ulong)((GoRuntimeFunc)f).Pc;          // returns uintptr
 
     public static long GOMAXPROCS(long n) => System.Environment.ProcessorCount;
     public static long NumCPU() => System.Environment.ProcessorCount;
